+++ src/main/java/com/poly/java5/Repository/PromotionRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 
 import com.poly.java5.Entity.Promotion;
 
-public interface PromotionRepository extends JpaRepository<Promotion, Long> {
+public interface PromotionRepository extends JpaRepository<Promotion, Integer> {
 
     
     @Query("SELECT p FROM Promotion p " +
@@ -21,7 +21,7 @@ public interface PromotionRepository extends JpaRepository<Promotion, Long> {
     	       "WHERE p.status = true " +
     	       "AND CURRENT_DATE BETWEEN p.startDate AND p.endDate " +
     	       "AND EXISTS (SELECT 1 FROM p.categories c WHERE c.id = :categoryId)")
-    	List<Promotion> findActivePromotionsByCategoryId(@Param("categoryId") Long categoryId);
+    	List<Promotion> findActivePromotionsByCategoryId(@Param("categoryId") Integer categoryId);
 
     	@Query("SELECT p FROM Promotion p " +
     	       "WHERE p.status = true " +
